@@ -1,10 +1,9 @@
 package com.javaweb.model.dto;
 
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
 
 import java.util.Collection;
 
@@ -13,6 +12,7 @@ import java.util.Collection;
 public class MyUserDetail extends User {
     private Long id;
     private String fullName;
+
 
     public MyUserDetail(String username, String password, boolean enabled, boolean accountNonExpired,
                         boolean credentialsNonExpired, boolean accountNonLocked,
